테마별 색 조회

+++ src/main/java/com/example/moro/app/post/repository/PostColorRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface PostColorRepository extends JpaRepository<PostColor, Long> {
+public interface PostColorRepository extends JpaRepository<PostColor,Integer> {
     // 게시물에 연결된 모든 후보 색상 id 가져오기
     @Query("SELECT pc.colormap.colorId From PostColor pc WHERE pc.post.id = :postId")
     List<Long> findColorIdsByPostId(@Param("postId") Long postId);
